+++ src/index.ts
@@ -1,1 +1,2 @@
-export * from './Tooltip.js';
+import './styles/main.css';
+export * from './Tooltip';